feat(exercises): correct streak cap
As a user, I want to be able to cap the correct streak of my exercises so that
I don't end up with exercises that have a correct streak of 1000 and will not
be shown anymore. Once a certain correct streak is reached it may really not
make sense anymore to increase it.

closes #35