refactor: removed query params in exercises page
Previously, we had query params in the exercise page
to reflect the current exercise page.
This did only work in one direction: They were
updated when the exercise filter changed but you
were not able to update the exercise filter by
changing the url's query params. Instead of making
this work, I removed this feature because this is
a "circular data flow" as I call it and error prone.
What I mean by this is that the url influences the
exercise filter and vice versa. It is difficult to
make this work properly so I deferred it.